requirements.txt
openpyxl==3.1.5
pandas==2.2.3
PyYAML==6.0.2
setuptools==65.5.0
